Move terminal related files to its own folder.
