+++ server/src/terminal/terminal.js
